+++ source/features/enable-file-links-in-compare-view.tsx
